Optimize part2
Use the insight that min location of a span of entries mapping all the
way to a location will be the first entry. The seed range can get
split as a you step through the mappings, so we find the minimum span
across all the mappings and use it to step through the seed range
instead of stepping one by one which is too expensive.

This version is way, way, way faster.